LPS-80122 (portal-tools-service-builder) Do not rewrite extending UADAnonymizer class

+++ modules/util/portal-tools-service-builder/src/main/java/com/liferay/portal/tools/service/builder/ServiceBuilder.java
@@ -4056,8 +4056,10 @@ public class ServiceBuilder {
 				_uadOutputPath, "/uad/anonymizer/", entity.getName(),
 				"UADAnonymizer.java"));
 
-		ToolsUtil.writeFile(
-			file, content, _author, _jalopySettings, _modifiedFileNames);
+		if (!file.exists()) {
+			ToolsUtil.writeFile(
+				file, content, _author, _jalopySettings, _modifiedFileNames);
+		}
 	}
 
 	private void _createUADAnonymizerTest(Entity entity) throws Exception {
